src/utils/jwt_util.py
from datetime import datetime, timedelta, timezone

import jwt
from jwt import ExpiredSignatureError, InvalidTokenError

from config.config import settings
from domain.errors.user_errors import InvalidCredentials


def create_access_token(data: dict):
    expire = datetime.now(timezone.utc) + timedelta(minutes=settings.JWT_EXPIRE_MINUTES)
    payload = {**data, "exp": expire}
    token = jwt.encode(payload, settings.JWT_SECRET, algorithm=settings.JWT_ALG)
    return token


def decode_token(token: str):
    try:
        payload = jwt.decode(token, settings.JWT_SECRET, algorithms=[settings.JWT_ALG])
    except ExpiredSignatureError:
        raise InvalidCredentials("Token expired")
    except InvalidTokenError:
        raise InvalidCredentials("Invalid token")
    return payload
