pyproject.toml
[project]
name = "finance-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "argon2-cffi>=25.1.0",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
]
